separate tests by function to improve readability and runtime efficiency during debugging
